+++ NEWS.md
@@ -1,5 +1,7 @@
 # googleCloudRunner 0.2.0.9000
 
+* Check for Cloud Scheduler Service Agent is present for scheduler to work (#73)
+
 # googleCloudRunner 0.2.0
 
 * Add `port` argument to Cloud Run deployments via `cr_buildstep_run()`